Reintroduced required layout.tsx to get nonce. Refactored.
