+++ frontend/src/app/msalInstance.ts
