+++ frontend/src/app/page.tsx
@@ -1,0 +1,6 @@
+import { VariantList } from "./variant";
+
+export default function Page() {
+
+    return <VariantList />
+}
