--- frontend/src/pages/index.tsx
@@ -1,9 +1,0 @@
-import { VariantList } from "@/app/variant";
-const Home = () => {
-
-    return (
-        <VariantList />
-    );
-}
-
-export default Home;
